close the compilation window if no error
I added a after-save hook to the golint to run it automatically. The finish hook will close the golint compilation window if no error, so you don't have to manually close it. This is useful if you already have before-save hook to gofmt, then it automatically runs golint against your saved file and disappears if there's no lint error.
Also your golint should only work if you are in the go mode.